Fading door fix
This script blocks any props that fall into the Fading Door zone, this is due to the fact that if you put a lot of props and activate the Fading Door on each one, the server will be overloaded and killed. Kids know about this feature and often use it to spite the owner.